+++ src/main/java/uk/ac/wlv/sentistrength/core/component/Paragraph.java
@@ -1,6 +1,11 @@
-package uk.ac.wlv.sentistrength;
+package uk.ac.wlv.sentistrength.core.component;
 
 import java.util.Random;
+
+import uk.ac.wlv.sentistrength.core.UnusedTermsClassificationIndex;
+import uk.ac.wlv.sentistrength.classification.ClassificationOptions;
+import uk.ac.wlv.sentistrength.classification.ClassificationResources;
+import uk.ac.wlv.sentistrength.option.TextParsingOptions;
 import uk.ac.wlv.utilities.Sort;
 import uk.ac.wlv.utilities.StringIndex;
 import uk.ac.wlv.wkaclass.Arff;
