+++ src/main/java/uk/ac/wlv/sentistrength/core/component/Sentence.java
@@ -3,8 +3,12 @@
 // (powered by Fernflower decompiler)
 //
 
-package uk.ac.wlv.sentistrength;
+package uk.ac.wlv.sentistrength.core.component;
 
+import uk.ac.wlv.sentistrength.core.UnusedTermsClassificationIndex;
+import uk.ac.wlv.sentistrength.classification.ClassificationOptions;
+import uk.ac.wlv.sentistrength.classification.ClassificationResources;
+import uk.ac.wlv.sentistrength.option.TextParsingOptions;
 import uk.ac.wlv.utilities.Sort;
 import uk.ac.wlv.utilities.StringIndex;
 import uk.ac.wlv.wkaclass.Arff;
