+++ src/main/java/uk/ac/wlv/sentistrength/wordlist/EmoticonsList.java
@@ -3,7 +3,7 @@
 // Decompiler options: packimports(3) fieldsfirst 
 // Source File Name:   EmoticonsList.java
 
-package uk.ac.wlv.sentistrength;
+package uk.ac.wlv.sentistrength.wordlist;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -13,6 +13,8 @@ import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
+
+import uk.ac.wlv.sentistrength.classification.ClassificationOptions;
 import uk.ac.wlv.utilities.FileOps;
 import uk.ac.wlv.utilities.Sort;
