+++ src/main/java/uk/ac/wlv/sentistrength/wordlist/EvaluativeTerms.java
@@ -3,7 +3,7 @@
 // Decompiler options: packimports(3) fieldsfirst
 // Source File Name:   EvaluativeTerms.java
 
-package uk.ac.wlv.sentistrength;
+package uk.ac.wlv.sentistrength.wordlist;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -13,6 +13,11 @@ import java.io.FileReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
+
+import uk.ac.wlv.sentistrength.classification.ClassificationOptions;
+import uk.ac.wlv.sentistrength.classification.ClassificationResources;
+import uk.ac.wlv.sentistrength.wordlist.IdiomList;
+import uk.ac.wlv.sentistrength.wordlist.SentimentWords;
 import uk.ac.wlv.utilities.FileOps;
 
 /**
