+++ src/main/java/uk/ac/wlv/sentistrength/wordlist/IdiomList.java
@@ -3,7 +3,7 @@
 // Decompiler options: packimports(3) fieldsfirst
 // Source File Name:   IdiomList.java
 
-package uk.ac.wlv.sentistrength;
+package uk.ac.wlv.sentistrength.wordlist;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -14,6 +14,9 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.util.Objects;
+
+import uk.ac.wlv.sentistrength.classification.ClassificationOptions;
+import uk.ac.wlv.sentistrength.classification.ClassificationResources;
 import uk.ac.wlv.utilities.FileOps;
 
 /**
@@ -42,7 +45,7 @@ public class IdiomList {
   /**
    * 处理习语中的单词时的下标。
    */
-  int[] igIdiomWordCount;
+  public int[] igIdiomWordCount;
 
   /**
    * 构造函数。
@@ -200,7 +203,6 @@ public class IdiomList {
    *
    * @param sPhrase 习语
    * @return 情感值
-   * @deprecated 弃用
    */
   public int getIdiomStrength_oldNotUseful(String sPhrase) {
     sPhrase = sPhrase.toLowerCase();
